Refactor and partial cleanup
